fix: reorganize files location
