+++ js/packages/components/inputs/chat/reply/interface.ts
@@ -1,0 +1,3 @@
+export interface ReplyMessageProps {
+	convPK: string
+}
