+++ js/packages/components/inputs/index.ts
@@ -3,3 +3,4 @@ export { MediumClearableInput } from './medium/MediumClearableInput'
 export { LargeInputWithIcon } from './large/LargeInputWithIcon'
 export { LargeInput } from './large/LargeInput'
 export { SmallInput } from './small/SmallInput'
+export { ChatTextInput } from './chat/ChatTextInput'
